intermediate/cpp/gpu/webgpu/webgpu-cross-platform-app/acceptance_tests/requirements.txt
PyWavelets==1.4.1
scipy==1.11.3
Pillow==10.0.1
imagehash==4.3.1
numpy==1.26.0
